requirements.txt
# Security-Log-Analyser — Dependencies

flask
python-dotenv
langchain-core
langchain-ollama
langgraph
rich